+++ src/AssemblyInfo.cs
@@ -1,5 +1,5 @@
 using System.Reflection;
 [assembly: AssemblyTitle("TouristTransfer")]
 [assembly: AssemblyDescription("Stock tourism contract crew transfers for KSP 1")]
-[assembly: AssemblyVersion("0.1.3.0")]
-[assembly: AssemblyFileVersion("0.1.3.0")]
+[assembly: AssemblyVersion("0.1.4.0")]
+[assembly: AssemblyFileVersion("0.1.4.0")]